+++ src/protocol/packet.cpp
@@ -33,9 +33,9 @@ Packet::SerializeInto(std::vector<uint8_t>& buf) const
     // Write header
     buf.push_back(static_cast<uint8_t>(type));
     // Write length in big endian
-    buf.push_back(ThirdByte(body.size()));
-    buf.push_back(SecondByte(body.size()));
-    buf.push_back(FirstByte(body.size()));
+    buf.push_back(ThirdByte((uint32_t)body.size()));
+    buf.push_back(SecondByte((uint32_t)body.size()));
+    buf.push_back(FirstByte((uint32_t)body.size()));
     // Write the body of the packet
     buf.insert(buf.end(), body.begin(), body.end());
 }
